requirements.txt
# Python Dependencies for Gas Station Analyzer
# 
# Production Dependencies (currently none - scripts use stdlib only)
# All production scripts use Python standard library only (urllib, json, os, sys, socket).

# Development Dependencies (Testing)
pytest>=7.4.0              # Testing framework
pytest-cov>=4.1.0          # Coverage reporting
pydantic>=2.0.0            # Data validation

# Future Production Dependencies (if needed):
python-dotenv>=1.0.0     # For better .env file handling
requests>=2.31.0         # For HTTP requests (alternative to urllib)


openpyxl